move feature classes to registry package
